Complete logging feature. Move some getter & setter from SenderProtocol to DataContainer. Code cleaning
